fix(tests): add negative parsing test for json to ensure our small ant fixture is accurate
